Consolidate Status > Resources page into a shared web-common component
Extract the resources filterable table + parse errors section into
`web-common/src/features/preview-mode/StatusResourcesPage.svelte`.
Both `/status/resources` and `/-/edit/status/resources` now thin-wrap
it. Drops the local-only `ParseErrorsSection` shim.
