+++ web-common/src/features/preview-mode/StatusResourcesPage.svelte
@@ -1,0 +1,123 @@
+<script lang="ts">
+  import { page } from "$app/stores";
+  import { onMount } from "svelte";
+  import {
+    ResourceKind,
+    SingletonProjectParserName,
+  } from "@rilldata/web-common/features/entity-management/resource-selectors";
+  import ParseErrorsSection from "@rilldata/web-common/features/resources/ParseErrorsSection.svelte";
+  import ResourcesFilterableTable from "@rilldata/web-common/features/resources/ResourcesFilterableTable.svelte";
+  import {
+    createRuntimeServiceCreateTriggerMutation,
+    createRuntimeServiceGetResource,
+    createRuntimeServiceListResources,
+    getRuntimeServiceListResourcesQueryKey,
+    V1ReconcileStatus,
+  } from "@rilldata/web-common/runtime-client";
+  import { useRuntimeClient } from "@rilldata/web-common/runtime-client/v2";
+  import {
+    createUrlFilterSync,
+    parseArrayParam,
+    parseStringParam,
+  } from "@rilldata/web-common/lib/url-filter-sync";
+  import { useQueryClient } from "@tanstack/svelte-query";
+
+  const runtimeClient = useRuntimeClient();
+  const queryClient = useQueryClient();
+  const createTrigger =
+    createRuntimeServiceCreateTriggerMutation(runtimeClient);
+
+  const filterSync = createUrlFilterSync([
+    { key: "kind", type: "array" },
+    { key: "status", type: "array" },
+    { key: "q", type: "string" },
+  ]);
+  filterSync.init($page.url);
+
+  let searchText = parseStringParam($page.url.searchParams.get("q"));
+  let selectedTypes = parseArrayParam($page.url.searchParams.get("kind"));
+  let selectedStatuses = parseArrayParam($page.url.searchParams.get("status"));
+  let mounted = false;
+
+  $: if (mounted && filterSync.hasExternalNavigation($page.url)) {
+    filterSync.markSynced($page.url);
+    selectedTypes = parseArrayParam($page.url.searchParams.get("kind"));
+    selectedStatuses = parseArrayParam($page.url.searchParams.get("status"));
+    searchText = parseStringParam($page.url.searchParams.get("q"));
+  }
+
+  $: if (mounted) {
+    filterSync.syncToUrl({
+      kind: selectedTypes,
+      status: selectedStatuses,
+      q: searchText,
+    });
+  }
+
+  onMount(() => {
+    mounted = true;
+  });
+
+  $: resourcesQuery = createRuntimeServiceListResources(
+    runtimeClient,
+    {},
+    { query: { refetchInterval: 5000 } },
+  );
+  $: resources = $resourcesQuery.data?.resources ?? [];
+
+  $: hasReconcilingSourcesOrModels = resources.some(
+    (r) =>
+      (r.meta?.name?.kind === ResourceKind.Source ||
+        r.meta?.name?.kind === ResourceKind.Model) &&
+      (r.meta?.reconcileStatus === V1ReconcileStatus.RECONCILE_STATUS_PENDING ||
+        r.meta?.reconcileStatus === V1ReconcileStatus.RECONCILE_STATUS_RUNNING),
+  );
+
+  $: projectParserQuery = createRuntimeServiceGetResource(
+    runtimeClient,
+    {
+      name: {
+        kind: ResourceKind.ProjectParser,
+        name: SingletonProjectParserName,
+      },
+    },
+    { query: { refetchOnMount: true, refetchOnWindowFocus: true } },
+  );
+  $: parseErrors =
+    $projectParserQuery.data?.resource?.projectParser?.state?.parseErrors ?? [];
+  $: parserReconcileError =
+    $projectParserQuery.data?.resource?.meta?.reconcileError;
+
+  function refreshAllSourcesAndModels() {
+    void $createTrigger.mutateAsync({ all: true }).then(() => {
+      void queryClient.invalidateQueries({
+        queryKey: getRuntimeServiceListResourcesQueryKey(
+          runtimeClient.instanceId,
+          undefined,
+        ),
+      });
+    });
+  }
+</script>
+
+<ResourcesFilterableTable
+  {resources}
+  containerHeight={550}
+  isLoading={$resourcesQuery.isLoading}
+  isError={$resourcesQuery.isError}
+  errorMessage={$resourcesQuery.error?.message ?? ""}
+  isRefreshDisabled={hasReconcilingSourcesOrModels}
+  onRefreshAll={refreshAllSourcesAndModels}
+  onRefetch={() => $resourcesQuery.refetch()}
+  bind:selectedStatuses
+  bind:selectedTypes
+  bind:searchText
+/>
+
+<ParseErrorsSection
+  {parseErrors}
+  {parserReconcileError}
+  isLoading={$projectParserQuery.isLoading}
+  isError={$projectParserQuery.isError}
+  errorMessage={$projectParserQuery.error?.message ?? ""}
+/>
